requirements.txt
openai
tavily-python
langchain
langchain-openai
grpcio
grpcio-tools
dotenv
colorama